some improvements in progress

+++ tap/extensions/http/main.go
@@ -237,7 +237,7 @@ func (d dissecting) Analyze(item *api.OutputChannelItem, resolvedSource string, 
 	}
 
 	if resDetails["bodySize"].(float64) < 0 {
-		resDetails["bodySize"] = 0
+		resDetails["bodySize"] = float64(0)
 	}
 
 	if item.Protocol.Version == "2.0" && !isRequestUpgradedH2C {
